refactor: rethink app architecture
Move canvas domain logic (OSML serializer, editor ops, node utils,
element configs) from app/components/canvas/ into lib/canvas/,
eliminating every lib -> app import. Enforce the boundary with an
ESLint no-restricted-imports rule so it cannot regress.
