+++ app/components/canvas/config/elementConfigs.ts
@@ -1,0 +1,3 @@
+// Compat re-export: keeps files touched by open PR #263 unchanged.
+// Fold into direct @/lib/canvas/elementConfigs imports once it lands.
+export type { AttributeSpec } from "@/lib/canvas/elementConfigs";
