+++ app/components/canvas/elements/ElementContainer.tsx
@@ -3,8 +3,8 @@ import { RenderElementProps, ReactEditor, useSlateStatic } from "slate-react";
 import { Node } from "slate";
 import type { CanvasContentElement, SceneElement } from "@/lib/canvas/types";
 import { isSceneElement } from "@/lib/canvas/scenes";
-import { ZERO_WIDTH_SPACE } from "../config/constants";
-import { ELEMENT_CONFIGS } from "../config/elementConfigs";
+import { ZERO_WIDTH_SPACE } from "@/lib/canvas/constants";
+import { ELEMENT_CONFIGS } from "@/lib/canvas/elementConfigs";
 import { useViewMode } from "../ViewModeContext";
 import { OutputPreview } from "./OutputPreview";
 import { DeleteButton } from "./DeleteButton";
