+++ app/components/canvas/elements/ModelBadge.tsx
@@ -2,7 +2,10 @@ import { useMemo } from "react";
 import { useConfig } from "@/lib/config/ConfigProvider";
 import type { ProviderKey } from "@/lib/connectors/types";
 import type { CanvasContentElement } from "@/lib/canvas/types";
-import { ELEMENT_CONFIGS, type AttributeSpec } from "../config/elementConfigs";
+import {
+	ELEMENT_CONFIGS,
+	type AttributeSpec,
+} from "@/lib/canvas/elementConfigs";
 import { AttributeBadge } from "./AttributeBadge";
 
 export function ModelBadge({ element }: { element: CanvasContentElement }) {
