+++ app/components/canvas/plugins/withNodeId.ts
@@ -1,6 +1,10 @@
 import { ReactEditor } from "slate-react";
 import type { CanvasEditor } from "@/lib/canvas/types";
-import { assignIdRecursively, makeNodeId, stripIds } from "../utils/nodeUtils";
+import {
+	assignIdRecursively,
+	makeNodeId,
+	stripIds,
+} from "@/lib/canvas/nodeUtils";
 
 export const withNodeId = (editor: ReactEditor): CanvasEditor => {
 	const { apply, insertFragment } = editor;
