+++ app/components/canvas/utils/buildGenerationJob.ts
@@ -3,7 +3,7 @@ import type { ProviderKey } from "@/lib/connectors/types";
 import { getDefaultConnector } from "@/lib/config/connectorUtils";
 import type { GenerationJob } from "@/lib/generation/queue";
 import type { CanvasContentElement } from "@/lib/canvas/types";
-import { ELEMENT_CONFIGS } from "../config/elementConfigs";
+import { ELEMENT_CONFIGS } from "@/lib/canvas/elementConfigs";
 
 export function buildGenerationJob(
 	element: CanvasContentElement,
