+++ lib/canvas/constants.ts
