+++ lib/canvas/guards.ts
