+++ lib/canvas/nodeUtils.ts
