+++ lib/canvas/osmlSerializer.ts
