+++ lib/canvas/parseXmlTag.ts
