+++ lib/canvas/useOSMLSerializer.ts
@@ -1,7 +1,7 @@
 import { useCallback, useRef, useState } from "react";
 import type { ParsedElement } from "@/lib/canvas/types";
 import { useConfig } from "@/lib/config/ConfigProvider";
-import { OSMLSerializer } from "../utils/osmlSerializer";
+import { OSMLSerializer } from "./osmlSerializer";
 
 const MAX_NODES_TO_SYNC = 3;
