+++ lib/generation/getPromptText.ts
@@ -1,7 +1,6 @@
 import { Node } from "slate";
 import type { CanvasContentElement } from "@/lib/canvas/types";
-
-const ZERO_WIDTH_SPACE = "​";
+import { ZERO_WIDTH_SPACE } from "@/lib/canvas/constants";
 
 export function getPromptText(element: CanvasContentElement): string {
 	return Node.string(element).replaceAll(ZERO_WIDTH_SPACE, "").trim();
